Fix delete mappings.

See merge request cse2115/2020-2010/7-student-house-food-management/op26-sem51!19

+++ requests/src/main/java/nl/tudelft/sem/requests/controllers/HouseController.java
@@ -17,8 +17,6 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-//import org.springframework.web.bind.annotation.RequestMapping;
-//import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -134,7 +132,12 @@ public class HouseController {
      */
     @DeleteMapping("/deleteHouse/{houseNumber}")
     public void deleteHouse(@PathVariable int houseNumber) {
-        houseRepository.deleteById(houseNumber);
+        Optional<House> house = houseRepository.findById(houseNumber);
+        if(house.isPresent()) {
+            houseRepository.deleteById(houseNumber);
+            System.out.println("house successfully deleted");
+        }
+        System.out.println("house not found!");
     }
 
     /**
